+++ tests/utils/test_qc_metrics.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 from pydantic import ValidationError
 
@@ -9,6 +10,8 @@ from BALSAMIC.utils.qc_metrics import (
     get_qc_metrics_dict,
     get_qc_available_panel_beds,
     merge_dicts,
+    get_multiqc_data_source,
+    extract_metrics_for_delivery,
 )
 
 
@@ -161,3 +164,91 @@ def test_get_qc_metrics_json_targeted(analysis_path):
             and "MEDIAN_TARGET_COVERAGE" in str(val_err)
             and "FOLD_80_BASE_PENALTY" in str(val_err)
         )
+
+
+def test_get_multiqc_data_source(analysis_path):
+    """test multiQC source extraction from multiqc_data.json analysis file"""
+
+    # GIVEN input parameters
+    sample = "concatenated_tumor_XXXXXX_R"
+    source_name_hs_metrics = "multiqc_picard_HsMetrics"
+    source_name_dup = "multiqc_picard_dups"
+
+    with open(
+        os.path.join(analysis_path, "qc", "multiqc_data", "multiqc_data.json"), "r"
+    ) as f:
+        raw_data = json.load(f)
+
+    # GIVEN an expected output
+    source_hs_metrics = "concatenated_tumor_XXXXXX_R.sorted.mrkdup.hsmetric"
+    source_dup = "concatenated_tumor_XXXXXX_R.sorted.mrkdup.txt"
+
+    # WHEN extracting the source of a specific sample and collection of metrics
+    out_source_hs_metrics = get_multiqc_data_source(
+        raw_data, sample, source_name_hs_metrics
+    )
+    out_source_dup = get_multiqc_data_source(raw_data, sample, source_name_dup)
+
+    # THEN check if the extracted source names correspond to the expected ones
+    assert source_hs_metrics == out_source_hs_metrics
+    assert source_dup == out_source_dup
+
+
+def test_extract_metrics_for_delivery(analysis_path):
+    """test output metrics retrieving"""
+
+    # GIVEN a sequencing type
+    seq_type = "targeted"
+
+    # GIVEN an expected output
+    n_metrics = 6  # Number of expected metric
+
+    hs_metric = {
+        "header": None,
+        "id": "tumor",
+        "input": "concatenated_tumor_XXXXXX_R.sorted.mrkdup.hsmetric",
+        "name": "PCT_OFF_BAIT",
+        "step": "multiqc_picard_HsMetrics",
+        "value": 0.364546,
+    }
+
+    ins_size_metric = {
+        "header": None,
+        "id": "tumor",
+        "input": "concatenated_tumor_XXXXXX_R.sorted.insertsizemetric",
+        "name": "MEAN_INSERT_SIZE",
+        "step": "multiqc_picard_insertSize",
+        "value": 201.813054,
+    }
+
+    dups_metric = {
+        "header": None,
+        "id": "tumor",
+        "input": "concatenated_tumor_XXXXXX_R.sorted.mrkdup.txt",
+        "name": "PERCENT_DUPLICATION",
+        "step": "multiqc_picard_dups",
+        "value": 0.391429,
+    }
+
+    # WHEN calling the function
+    metrics = extract_metrics_for_delivery(analysis_path, seq_type)
+
+    # THEN check if the metrics are correctly retrieved
+    assert len(metrics) == n_metrics
+    assert (
+        hs_metric in metrics and ins_size_metric in metrics and dups_metric in metrics
+    )
+
+
+def test_extract_metrics_for_delivery_filtering_umi(analysis_path):
+    """test umi discarding when extracting metrics"""
+
+    # GIVEN a sequencing type
+    seq_type = "targeted"
+
+    # WHEN calling the function
+    metrics = extract_metrics_for_delivery(analysis_path, seq_type)
+
+    # THEN check if the umi samples are filtered out
+    for metric in metrics:
+        assert "umi" not in metric["input"]